requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
joblib==1.2.0
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
typing-extensions==4.4.0
whitenoise==6.2.0
